seperated coding menu and realted methods from sleep menu and its related methods
